\name{gr95ModelA}
\alias{gr95ModelA}
\title{An atomic root finder to model a Greco 1995 surface
for two vectors of concentrations. Not currently used:
planned for surface plotting.}
\usage{
  gr95ModelA(params, data, mup)
}
\arguments{
  \item{mup}{A logical variable indicating whether the dose
  slopes are positive. Usually false.}

  \item{params}{A numeric vector, always of the form B,
  Econ, C50.1, C50.2, m1, m2, alpha.}

  \item{data}{A 2-entry vector, a point at which the
  response is to be modelled.}
}
\description{
  An atomic root finder to model a Greco 1995 surface for
  two vectors of concentrations. Not currently used:
  planned for surface plotting.
}
\author{
  Paul Lakin
}

